forteenth commmit

+++ frontend/src/components/users/beautician/services/Addservicemodal.js
@@ -6,8 +6,10 @@ import Modal from "@mui/material/Modal";
 import { useState, useEffect } from "react";
 import TextField from "@mui/material/TextField";
 import { useDispatch, useSelector } from "react-redux";
-import { setAllservices,setServices } from "../../../../feautures/loginslice";
+import { setAllservices, setServices } from "../../../../feautures/loginslice";
 import axiosInstance from "../../../../axios/axiosconfig";
+import Input from "@mui/joy/Input";
+import MuiAlert from "@mui/material/Alert";
 
 const style = {
   position: "absolute",
@@ -21,43 +23,58 @@ const style = {
   p: 4,
 };
 
+const Alert = React.forwardRef(function Alert(props, ref) {
+  return <MuiAlert elevation={6} ref={ref} variant="filled" {...props} />;
+});
+
 const Addservicemodal = () => {
   const dispatch = useDispatch();
   const statedatas = useSelector((state) => state.login);
   const [selectedService, setSelectedService] = useState("");
+  const [addServiceError, setAddServiceError] = useState(false);
+  const [serviceFeeError, setServiceFeeError] = useState(false);
+
+  const [serviceFee, setServiceFee] = useState("");
   const [open, setOpen] = useState(false);
-  const [changed,setChanged]=useState(false)
+  const [changed, setChanged] = useState(false);
   const handleOpen = () => {
     setOpen(true);
   };
   const handleClose = () => {
     setOpen(false);
+    setAddServiceError(false);
+    setServiceFeeError(false);
   };
 
   useEffect(() => {
     const allservices = localStorage.getItem("allservices-B");
-    const services=localStorage.getItem("services-B");
+    const services = localStorage.getItem("services-B");
     const parsedallservices = JSON.parse(allservices);
     const parsedservices = JSON.parse(services);
 
     dispatch(setAllservices(parsedallservices));
     dispatch(setServices(parsedservices));
-  }, [changed,]);
+  }, [changed]);
 
   const handleAddService = () => {
     const datas = {
       servicename: selectedService,
       beautid: statedatas.value.beautdetails.id,
+      servicefee: serviceFee,
     };
+    console.log(datas, "*****CURRENT******");
     axiosInstance
       .post("beaut/addnewservice/", datas)
-      .then((res) =>{
-        localStorage.setItem("services-B",JSON.stringify(res.data.services))
-        setChanged((prev)=>!prev)
-        handleClose()
-
-
-      } )
+      .then((res) => {
+        if (res.data.message == "Already Present") {
+          setAddServiceError(true);
+         
+        } else {
+          localStorage.setItem("services-B", JSON.stringify(res.data.services));
+          setChanged((prev) => !prev);
+          handleClose();
+        }
+      })
       .catch((error) => alert(error));
   };
   return (
@@ -95,12 +112,37 @@ const Addservicemodal = () => {
                 id="mySelect"
                 name="fruit"
                 className="form-control"
-                onChange={(e) => setSelectedService(e.target.value)}
+                onChange={(e) => {
+                  setSelectedService(e.target.value);
+                  console.log(selectedService, "*********************");
+                }}
               >
                 {statedatas.value.allservices.map((item) => {
                   return <option>{item.name}</option>;
                 })}
               </select>
+              <label className="mt-4">Fee</label>
+              <Input
+                className="form-control"
+                size="md"
+                placeholder="Enter Fee"
+                onChange={(e) =>{
+                    const val=e.target.value
+                    const regex = /^[0-9]+$/;   
+                    if (regex.test(val)){
+                        setServiceFee(e.target.value)
+                        setServiceFeeError(false)
+
+                    }
+                    else{
+                        setServiceFeeError(true)
+                    }
+                } }
+              />
+              {
+                serviceFeeError&& 
+                <span className="text-danger">Please Enter Valid Price!</span>
+              }
             </div>
             <div className="row">
               <Button
@@ -125,6 +167,17 @@ const Addservicemodal = () => {
               >
                 Close
               </Button>
+              {addServiceError && (
+                <Alert
+                  severity="error"
+                  sx={{
+                    marginTop: "20px",
+                    marginLeft: "70px",
+                  }}
+                >
+                  Service Already Added!
+                </Alert>
+              )}
             </div>
           </Typography>
         </Box>
